More code!
Custom WE schematic(MCEditExtened), that saves material name instead of
IDs for cross-server compatibility. //schematic save -f mcee <filename>.
More work done on smoothing out the Manager classes.
Instances can now be mounted and unmounted, and use AsynchWorldEdit to
make it happen (this part *may* need some work, but it does in fact
work).
Started on "Dungeon Edit Mode", to allow admins to modify the schematic
data prior to being finalized and used for Instances.
Changed all the boolean returning functions with RetVal, so messages can
be passed to calling functions, along with status, in a non-stupid way.
Added a Log class, as a bridge between server-tech agnostic main
classes, and the bukkit plugin. Trying to keep the backend code as
flexible as I can to make porting easier.