Async Pasting, and some code cleanup.
Added AsyncManager - Might be hacky, and/or break with multiple
instances being pasted in - but for now, it works.
Had to modify the flow a bit in the other Managers - they now wait until
the AsyncManager informs them the paste is done before finishing
instance creation.
Cleaned up the Help - made it a bit less spammy, and more uniform.
Dungeons now get pasted in Asynchronously - yay. Complements the Asynch
removal nicely ;)
Edit Instance changes are now temp-stored. When the dungeon is
finalized, the pre-edit version is saved off, and the schematic is
replaced with the changes you made!